change to binary_search_by_recursion
Current version of binary_search_by_recursion does not support values that do not exist. For example, this input will return a recursion error:

collection = [1,2,3,4,5]
print(binary_search(collection, 7, 0, len(collection)-1))

To fix, I added this code, which checks to make sure that the item is in range. This code is similar to what is already present for the non-recursive function in this file:

    if (right < left):
        return None 

I also swapped these lines in the same function to make the code easier to read. The new code also matches what is used in the non-recursive function.

old: midpoint = left + (right - left) // 2
new: midpoint = (left + right) // 2